+++ runelite-plugin.properties
@@ -3,4 +3,4 @@ author=BrastaSauce, wesley221
 support=
 description=Displays best location and information for Turael/Aya tasks
 tags=turael,skipping,slayer,boosting,tureal,aya
-plugins=com.brastasauce.turaelskipping.TuraelSkippingPlugin
+plugins=com.mathiaslj.configurableslayerinfo.TuraelSkippingPlugin
