+++ src/main/java/com/mathiaslj/configurableslayerinfo/models/SlayerTask.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.brastasauce.turaelskipping.models;
+package com.mathiaslj.configurableslayerinfo.models;
 
 import lombok.Getter;
 import net.runelite.api.coords.WorldPoint;
